refactor: improve size and clarify component
