+++ packages/frontend/src/features/sqlRunner/components/ContentPanel.tsx
@@ -59,7 +59,7 @@ import {
 } from '../store/sqlRunnerSlice';
 import { DownloadCsvButton } from './DownloadCsvButton';
 import { SqlEditor, type MonacoHighlightChar } from './SqlEditor';
-import { SqlQueryHistorySelect } from './SqlQueryHistorySelect';
+import { SqlQueryHistory } from './SqlQueryHistory';
 
 const DEFAULT_SQL_LIMIT = 500;
 
@@ -361,7 +361,7 @@ export const ContentPanel: FC = () => {
                         </Group>
                         <Group spacing="xs">
                             {activeEditorTab === EditorTabs.SQL && (
-                                <SqlQueryHistorySelect />
+                                <SqlQueryHistory />
                             )}
                             <RunSqlQueryButton
                                 isLoading={isLoading}
